Se ha componetizado la lista de productos.
Bug: En Favoritos no quiere mostrar la pagina 3
