+++ pages/components/header.js
@@ -11,7 +11,7 @@ export default function Header() {
             <Link href="/"><a><Image src="/icon.svg" alt="Vercel Logo" width={161} height={30}/></a></Link>
            </div>
            <div class="favorito">
-           <Link href="/favoritos"><a><Image src="/corazon.svg" alt="Search icon" width={25} height={25} objectFit='contain'></Image></a></Link>
+           <a href="/favoritos"><Image src="/corazon.svg" alt="Search icon" width={25} height={25} objectFit='contain'></Image></a>
             </div>
             <div class="linea2"><hr></hr></div>
             <Search/>
