+++ styles/globals.css
@@ -168,9 +168,10 @@ footer a:hover:not(.active) {background-color: #ddd;}
   height: 180px;
 }
 .producto{
-  margin-left:9em;
+  margin-left:7.5%;
   margin-top:3em;
   position: relative;
+  text-align: left;
 }
 .producto img{
   background-color: rgb(245, 245, 245);
